# docs/openapi.json
{
  "components" : {
    "schemas" : {
      "ArticleCommentCreateRequest" : {
        "properties" : {
          "content" : {
            "maxLength" : 5000,
            "minLength" : 0,
            "type" : "string"
          },
          "nickname" : {
            "maxLength" : 50,
            "minLength" : 0,
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "password" : {
            "maxLength" : 100,
            "minLength" : 4,
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ArticleCommentListItem" : {
        "properties" : {
          "content" : {
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "maskedIp" : {
            "type" : "string"
          },
          "nickname" : {
            "type" : "string"
          },
          "parentId" : {
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "ArticleCommentListResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ArticleCommentListItem"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
      "ArticleDetailResponse" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "commentCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "id" : {
            "format" : "int64",
            "type" : "integer"
          },
          "likeCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          },
          "sourceUrl" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "thumbnailUrl" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "ArticleLikeStateResponse" : {
        "properties" : {
          "articleId" : {
            "format" : "int64",
            "type" : "integer"
          },
          "likeCount" : {
            "format" : "int64",
            "type" : "integer"
          },
          "likedByReader" : {
            "type" : "boolean"
          }
        },
        "type" : "object"
      },
      "ArticleListItem" : {
        "properties" : {
          "category" : {
            "type" : "string"
          },
          "id" : {
            "type" : "string"
          },
          "publishedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "source" : {
            "type" : "string"
          },
          "summary" : {
            "type" : "string"
          },
          "thumbnailUrl" : {
            "type" : [ "string", "null" ]
          },
          "title" : {
            "type" : "string"
          }
        },
        "required" : [ "category", "id", "publishedAt", "source", "summary", "thumbnailUrl", "title" ],
        "type" : "object"
      },
      "ArticleListResponse" : {
        "properties" : {
          "items" : {
            "items" : {
              "$ref" : "#/components/schemas/ArticleListItem"
            },
            "type" : "array"
          },
          "nextCursor" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
      "CategoryListItem" : {
        "properties" : {
          "slug" : {
            "type" : "string"
          },
          "sortOrder" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      }
    }
  },
  "info" : {
    "description" : "Public API for the everytldr reader experience.",
    "title" : "everytldr API",
    "version" : "v0"
  },
  "openapi" : "3.1.0",
  "paths" : {
    "/api/articles" : {
      "get" : {
        "operationId" : "listArticles",
        "parameters" : [ {
          "in" : "query",
          "name" : "cursor",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "categoryPrefix",
          "required" : false,
          "schema" : {
            "type" : "string"
          }
        }, {
          "in" : "query",
          "name" : "size",
          "required" : false,
          "schema" : {
            "format" : "int32",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleListResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      }
    },
    "/api/articles/{articleId}/comments" : {
      "get" : {
        "operationId" : "listArticleComments",
        "parameters" : [ {
          "in" : "path",
          "name" : "articleId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleCommentListResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      },
      "post" : {
        "operationId" : "createArticleComment",
        "parameters" : [ {
          "in" : "path",
          "name" : "articleId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/ArticleCommentCreateRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleCommentListItem"
                }
              }
            },
            "description" : "Created"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      }
    },
    "/api/articles/{articleId}/likes/me" : {
      "delete" : {
        "operationId" : "unlikeArticle",
        "parameters" : [ {
          "in" : "path",
          "name" : "articleId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleLikeStateResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      },
      "get" : {
        "operationId" : "getMyArticleLike",
        "parameters" : [ {
          "in" : "path",
          "name" : "articleId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleLikeStateResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      },
      "put" : {
        "operationId" : "likeArticle",
        "parameters" : [ {
          "in" : "path",
          "name" : "articleId",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleLikeStateResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      }
    },
    "/api/articles/{id}" : {
      "get" : {
        "operationId" : "getArticle",
        "parameters" : [ {
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ArticleDetailResponse"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Articles" ]
      }
    },
    "/api/categories" : {
      "get" : {
        "operationId" : "listCategories",
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/CategoryListItem"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "OK"
          },
          "400" : {
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "Not Found"
          },
          "503" : {
            "description" : "Service Unavailable"
          }
        },
        "tags" : [ "Categories" ]
      }
    }
  },
  "servers" : [ {
    "description" : "Generated server url",
    "url" : "http://localhost"
  } ]
}